[fix]: corregir bugs detectados en testing E2E (ajustes manuales de stock y caja, ajuste por proveedor, resumen SQLite)
RegistrarMovimientoAsync (stock) ya no adjunta navegaciones AsNoTracking que se re-insertaban (UNIQUE constraint). RegistrarMovimientoAsync (caja) asigna Id_usuario con fallback a quien abrio la caja (FOREIGN KEY). Ajuste de precios por proveedor usa una consulta sin navegaciones para no duplicar Categoria. Resumen de periodo traduce Sum(decimal) a double para SQLite. +5 tests de regresion discriminativos (433 en verde).

+++ GestionComercial.Persistencia/Repositorio/MovimientoStockRepositorio.cs
@@ -137,7 +137,9 @@ namespace GestionComercial.Persistencia.Repositorio
                 {
                     TipoMovimiento = g.Key,
                     Cantidad = g.Count(),
-                    Unidades = g.Sum(m => m.Cantidad)
+                    // SQLite no traduce Sum sobre decimal: se suma como double (cast explícito)
+                    // y se convierte a int al mapear el DTO (líneas de abajo).
+                    Unidades = g.Sum(m => (double)m.Cantidad)
                 })
                 .ToListAsync(ct);
